CET-5850   [Backstage] Respects tenant end date
