+++ src/components/Entitlements/ExpirationBanner.tsx
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 import React from 'react';
-import { ExpirationResponse} from "../../api/types";
-import {WarningPanel} from "@backstage/core-components";
-import {isNil} from "lodash";
-import {maybePluralize} from "../../utils/strings";
+import { ExpirationResponse } from "../../api/types";
+import { WarningPanel } from "@backstage/core-components";
+import { isNil } from "lodash";
 import moment from "moment";
+import { getExpirationMessage } from "./ExpirationUtils";
 
-interface ExpirationBannerProps extends ExpirationResponse {}
+interface ExpirationBannerProps extends ExpirationResponse {
+}
 
 export const ExpirationBanner = ({ contractType, expirationDate, shutdownDate }: ExpirationBannerProps) => {
   const isTrial = contractType === 'TRIAL';
@@ -46,47 +47,3 @@ export const ExpirationBanner = ({ contractType, expirationDate, shutdownDate }:
   );
 };
 
-
-export const getExpirationMessage = (
-  isTrial: boolean,
-  daysUntilExpiration: number,
-  daysPastExpiration?: number,
-  daysUntilShutdown?: number
-) => {
-  return isTrial
-    ? getTrialExpirationMessage(daysUntilExpiration)
-    : getProdExpirationMessage(daysUntilExpiration, daysPastExpiration, daysUntilShutdown);
-};
-
-const getProdExpirationMessage = (
-  daysUntilExpiration: number,
-  daysPastExpiration?: number,
-  daysUntilShutdown?: number
-) => {
-  const shutdownDates =
-    !isNil(daysUntilShutdown) && daysUntilShutdown > 0
-      ? `in ${maybePluralize(daysUntilShutdown, 'day')}`
-      : 'today';
-
-  const shutdownMessage =
-    !isNil(daysUntilShutdown)
-      ? `. Your workspace will automatically shut down ${shutdownDates}`
-      : '';
-
-  const expirationState =
-    daysUntilExpiration > 0
-      ? `is expiring in ${maybePluralize(daysUntilExpiration, 'day')}`
-      : daysUntilExpiration === 0
-        ? `expires today${shutdownMessage}`
-        : `expired ${daysPastExpiration} days ago${shutdownMessage}`;
-
-  return `Your access ${expirationState}. Please reach out to your account manager for support.`;
-};
-
-const getTrialExpirationMessage = (daysUntilExpiration: number) => {
-  return daysUntilExpiration > 0
-    ? `Your Cortex trial expires in ${maybePluralize(daysUntilExpiration, 'day')}.`
-    : daysUntilExpiration === 0
-      ? 'Your Cortex trial expires today - Please contact your account manager to continue using Cortex!'
-      : 'Your Cortex trial has expired - Please contact your account manager to continue using Cortex!';
-};
